+++ packages/soda-hooks/CHANGELOG.md
@@ -1,5 +1,16 @@
 # soda-hooks
 
+## 6.2.0
+
+### Minor Changes
+
+- 修改 Excel 相关功能，新增 useSize Hook
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.17.0
+
 ## 6.1.21
 
 ### Patch Changes